refactor: add ubuntu support and fix alpine

+++ lxc/nginx-proxy-manager/setup.sh
@@ -1,13 +1,27 @@
 #!/usr/bin/env sh
-set -euo pipefail
-trap 'rm -rf $TMP' EXIT SIGINT SIGTERM
-
 TMP=/tmp/npm_install.sh
-URL=https://raw.githubusercontent.com/ej52/proxmox/main/lxc/nginx-proxy-manager/install.sh
+URL=https://raw.githubusercontent.com/ej52/proxmox/main/lxc/nginx-proxy-manager/install
+
+if [ "$(uname)" != "Linux" ]; then
+  echo "OS NOT SUPPORTED"
+  exit 1
+fi
+
+DISTRO=$(cat /etc/*-release | grep -w ID | cut -d= -f2 | tr -d '"')
+if [ "$DISTRO" != "alpine" ] && [ "$DISTRO" != "ubuntu" ]; then
+  echo "DISTRO NOT SUPPORTED"
+  exit 1
+fi
 
 rm -rf $TMP
-wget -q -O "$TMP" "$URL"
+wget -O "$TMP" "$URL/$DISTRO.sh"
 
 chmod +x "$TMP"
-sh "$TMP"
+
+if [ "$(command -v bash)" ]; then
+  bash "$TMP"
+else
+  sh "$TMP"
+fi
+
